basestationproject/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
click==8.1.8
Django==5.1.4
djangorestframework==3.15.2
empy==4.2
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.1
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
